requirements.txt
aiomysql==0.2.0
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
et-xmlfile==1.1.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
mysql-connector-python==8.2.0
openpyxl==3.1.2
protobuf==4.21.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyMySQL==1.1.0
pyparsing==3.1.1
requests==2.31.0
rsa==4.9
soupsieve==2.5
uritemplate==4.1.1
urllib3==2.1.0
websockets==12.0
